refactor: 重构灵动岛窗口状态核心接口
